hashing_service/requirements.txt
fastapi
pydantic
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
prometheus-fastapi-instrumentator